refactor: move out flattendialog
